[REFACTORING] - Innfører interface for alle gateways.
Bruker interface for å styre avhengigheten, slik at alt peker inn mot domenet.
