+++ src/main/java/no/nav/fo/veilarbregistrering/orgenhet/HentEnheterGateway.java
@@ -1,0 +1,8 @@
+package no.nav.fo.veilarbregistrering.orgenhet;
+
+import java.util.List;
+
+public interface HentEnheterGateway {
+
+    List<NavEnhet> hentAlleEnheter();
+}
