+++ src/main/java/no/nav/fo/veilarbregistrering/registrering/bruker/db/BrukerRegistreringRepositoryImpl.java
@@ -2,7 +2,6 @@ package no.nav.fo.veilarbregistrering.registrering.bruker.db;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import no.nav.fo.veilarbregistrering.db.UtdanningUtils;
 import no.nav.fo.veilarbregistrering.registrering.bruker.AktorId;
 import no.nav.fo.veilarbregistrering.registrering.bruker.OrdinaerBrukerRegistrering;
 import no.nav.fo.veilarbregistrering.registrering.bruker.SykmeldtRegistrering;
