+++ src/test/java/no/nav/fo/veilarbregistrering/arbeidsforhold/resources/ArbeidsforholdResourceTest.java
@@ -2,7 +2,7 @@ package no.nav.fo.veilarbregistrering.arbeidsforhold.resources;
 
 import no.nav.apiapp.security.veilarbabac.VeilarbAbacPepClient;
 import no.nav.dialogarena.aktor.AktorService;
-import no.nav.fo.veilarbregistrering.arbeidsforhold.adapter.ArbeidsforholdGateway;
+import no.nav.fo.veilarbregistrering.arbeidsforhold.ArbeidsforholdGateway;
 import no.nav.fo.veilarbregistrering.bruker.UserService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
