+++ src/test/java/no/nav/fo/veilarbregistrering/config/ApplicationTestConfig.java
@@ -4,7 +4,7 @@ import no.nav.apiapp.config.ApiAppConfigurator;
 import no.nav.apiapp.security.veilarbabac.VeilarbAbacPepClient;
 import no.nav.dialogarena.aktor.AktorService;
 import no.nav.fo.veilarbregistrering.mock.*;
-import no.nav.fo.veilarbregistrering.arbeidsforhold.adapter.ArbeidsforholdGateway;
+import no.nav.fo.veilarbregistrering.arbeidsforhold.ArbeidsforholdGateway;
 import no.nav.fo.veilarbregistrering.bruker.UserService;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Conditional;
